# Produce debug entries for inlined methods