--- a/substratevm/src/com.oracle.objectfile/src/com/oracle/objectfile/debugentry/Range.java
+++ b/substratevm/src/com.oracle.objectfile/src/com/oracle/objectfile/debugentry/Range.java
@@ -123,7 +123,7 @@ public class Range {
     }
 
     public boolean isDeoptTarget() {
-        return methodEntry.isDeoptTarget;
+        return methodEntry.isDeopt();
     }
 
     private String getExtendedMethodName(boolean includeClass, boolean includeParams, boolean includeReturnType) {