## Known Issues

* Given two call-chains:
  1. foo->bar->baz
  2. foo->bar->buz 
we could create a "concrete inlined instance tree" (per the DWARF spec) like this:
```
   foo
    |
   bar
    |
   / \
  /   \
baz   buz
```
instead we generate two separate trees, resulting in more DIEs (Debug Information Entries).

* Ideally an `DW_TAG_inlined_subroutine` would reference (through ` DW_AT_abstract_origin`) the `DW_TAG_subprogram` defining the corresponding subprogram even if it's declared in a different compilation unit. There are cases however where when creating the `DW_TAG_inlined_subroutine` the corresponding `DW_TAG_subprogram` has not been created yet (even worse the corresponding compilation unit might not have been created yet). To overcome this issue our current approach creates partial `DW_TAG_subprogram` entries in the compilation unit where the inlining happens. This however comes at the cost of having more DIEs.

Before this PR hello.hello with debug symbols is 10MB
After this PR hello.hello with debug symbols becomes 38MB
Without debug symbols hello.hello is 7.9MB

Closes #2701