@zakkak 2020-09-30
Examples of DIEs referenced above:

### Partial `DW_TAG_subprogram`

```
 <1><38>: Abbrev Number: 5 (DW_TAG_subprogram)
    <39>   DW_AT_name        : (indirect string, offset: 0x34686): com.oracle.svm.core.os.ImageHeapProvider::get
    <3d>   DW_AT_inline      : 1        (inlined)
    <3e>   DW_AT_decl_file   : 0x3
    <42>   DW_AT_decl_line   : 0x2c
    <46>   DW_AT_external    : 1
```

### Complete `DW_TAG_subprogram`
```
 <1><47>: Abbrev Number: 4 (DW_TAG_subprogram)
    <48>   DW_AT_name        : (indirect string, offset: 0x6bae): com.oracle.svm.core.Isolates::create(org.graalvm.nativeimage.c.type.WordPointer, com.oracle.svm.core.c.function.CEntryPointCreateIsolateParameters)
    <4c>   DW_AT_low_pc      : 0x405000
    <54>   DW_AT_high_pc     : 0x405071
    <5c>   DW_AT_decl_file   : 0x1
    <60>   DW_AT_decl_line   : 0x48
    <64>   DW_AT_external    : 1
```

### Concrete inlined instance trees

```
 <1><137>: Abbrev Number: 4 (DW_TAG_subprogram)
    <138>   DW_AT_name        : (indirect string, offset: 0x1ad5e): com.oracle.svm.core.Isolates::tearDownCurrent
    <13c>   DW_AT_low_pc      : 0x405080
    <144>   DW_AT_high_pc     : 0x4050f9
    <14c>   DW_AT_decl_file   : 0x1
    <150>   DW_AT_decl_line   : 0x5e
    <154>   DW_AT_external    : 1
 <2><155>: Abbrev Number: 8 (DW_TAG_inlined_subroutine)
    <156>   DW_AT_abstract_origin: <0xa1>
    <15a>   DW_AT_low_pc      : 0x405084
    <162>   DW_AT_high_pc     : 0x40508f
    <16a>   DW_AT_call_file   : 0x1
    <16e>   DW_AT_call_line   : 0x5e
 <3><172>: Abbrev Number: 8 (DW_TAG_inlined_subroutine)
    <173>   DW_AT_abstract_origin: <0xb0>
    <177>   DW_AT_low_pc      : 0x405084
    <17f>   DW_AT_high_pc     : 0x40508f
    <187>   DW_AT_call_file   : 0x1
    <18b>   DW_AT_call_line   : 0x65
 <4><18f>: Abbrev Number: 7 (DW_TAG_inlined_subroutine)
    <190>   DW_AT_abstract_origin: <0xbf>
    <194>   DW_AT_low_pc      : 0x405084
    <19c>   DW_AT_high_pc     : 0x40508f
    <1a4>   DW_AT_call_file   : 0x4
    <1a8>   DW_AT_call_line   : 0xfa
 <4><1ac>: Abbrev Number: 0
 <3><1ad>: Abbrev Number: 0
 <2><1ae>: Abbrev Number: 8 (DW_TAG_inlined_subroutine)
    <1af>   DW_AT_abstract_origin: <0xa1>
    <1b3>   DW_AT_low_pc      : 0x40508f
    <1bb>   DW_AT_high_pc     : 0x405093
    <1c3>   DW_AT_call_file   : 0x1
    <1c7>   DW_AT_call_line   : 0x5e
 <3><1cb>: Abbrev Number: 8 (DW_TAG_inlined_subroutine)
    <1cc>   DW_AT_abstract_origin: <0xb0>
    <1d0>   DW_AT_low_pc      : 0x40508f
    <1d8>   DW_AT_high_pc     : 0x405093
    <1e0>   DW_AT_call_file   : 0x1
    <1e4>   DW_AT_call_line   : 0x65
 <4><1e8>: Abbrev Number: 7 (DW_TAG_inlined_subroutine)
    <1e9>   DW_AT_abstract_origin: <0xce>
    <1ed>   DW_AT_low_pc      : 0x40508f
    <1f5>   DW_AT_high_pc     : 0x405093
    <1fd>   DW_AT_call_file   : 0x4
    <201>   DW_AT_call_line   : 0xfa
 <4><205>: Abbrev Number: 0
 <3><206>: Abbrev Number: 0
```

Note that this could become:
```
 <1><137>: Abbrev Number: 4 (DW_TAG_subprogram)
    <138>   DW_AT_name        : (indirect string, offset: 0x1ad5e): com.oracle.svm.core.Isolates::tearDownCurrent
    <13c>   DW_AT_low_pc      : 0x405080
    <144>   DW_AT_high_pc     : 0x4050f9
    <14c>   DW_AT_decl_file   : 0x1
    <150>   DW_AT_decl_line   : 0x5e
    <154>   DW_AT_external    : 1
 <2><155>: Abbrev Number: 8 (DW_TAG_inlined_subroutine)
    <156>   DW_AT_abstract_origin: <0xa1>
    <15a>   DW_AT_low_pc      : 0x405084
    <162>   DW_AT_high_pc     : 0x405093
    <16a>   DW_AT_call_file   : 0x1
    <16e>   DW_AT_call_line   : 0x5e
 <3><172>: Abbrev Number: 8 (DW_TAG_inlined_subroutine)
    <173>   DW_AT_abstract_origin: <0xb0>
    <177>   DW_AT_low_pc      : 0x405084
    <17f>   DW_AT_high_pc     : 0x405093
    <187>   DW_AT_call_file   : 0x1
    <18b>   DW_AT_call_line   : 0x65
 <4><18f>: Abbrev Number: 7 (DW_TAG_inlined_subroutine)
    <190>   DW_AT_abstract_origin: <0xbf>
    <194>   DW_AT_low_pc      : 0x405084
    <19c>   DW_AT_high_pc     : 0x40508f
    <1a4>   DW_AT_call_file   : 0x4
    <1a8>   DW_AT_call_line   : 0xfa
 <4><1e8>: Abbrev Number: 7 (DW_TAG_inlined_subroutine)
    <1e9>   DW_AT_abstract_origin: <0xce>
    <1ed>   DW_AT_low_pc      : 0x40508f
    <1f5>   DW_AT_high_pc     : 0x405093
    <1fd>   DW_AT_call_file   : 0x4
    <201>   DW_AT_call_line   : 0xfa
 <4><1ac>: Abbrev Number: 0
 <3><1ad>: Abbrev Number: 0
```