@adinn 2020-10-01
@zakkak This looks very interesting but it's going to need to wait on some large, upcoming changes to the current debuginfo code that I have queued up.

I have almost completed adding types to the generated info model and that includes reorganizing the method debug info so that it is associated with class info records. It also involves following the gcc model of splitting it into separate declarations and definitions for the methods. Your changes will need to be reworked into that patch and I think it would be best to wait for it to arrive before doing so.