@zakkak 2021-03-16
Note: The current implementation uses "pseudo-method" entries for the nested inline methods. So instead of using `DW_AT_specification` to link from an inlined method location to the corresponding method declaration I use `DW_AT_name`, `DW_AT_decl_file` and `DW_AT_decl_line` to provide enough info for gdb to produce the backtrace.