@zakkak 2021-06-29
@adinn can you please review this?

The implementation associates ranges corresponding to inlined methods with their "caller" ranges and then using that info generates the corresponding concrete instance trees.

Note the implementation still generates more concrete instance trees than we actually need (see "Known Issues" in https://github.com/oracle/graal/pull/2880#issue-495187547). My understanding is that in order to solve this we would need to change the way we represent subranges, currently as a list of contiguous memory addresses, to a list of subtrees representing the inlined call trees. I would like to work on this on a different PR if you agree.

Regarding perfomance and debug info size please see the following up to date table:

  |                                    | `master`    | this branch  |
  |------------------------------------|-------------|--------------|
  | -H:+OmitInlinedMethodDebugLineInfo | ~82s / 14MB | ~82s / 14MB  |
  | -H:-OmitInlinedMethodDebugLineInfo | ~90s / 17MB | ~330s / 65MB |