@zakkak 2021-07-21
Thanks to @adinn who identified the main cause of the larger binaries being the fact that I was creating the abstract inline DIEs in the CU of the caller class instead of the CU of the class implementing the corresponding method. As a result instead of reusing the same abstract inline DIE I was replicating them in every class they were being referenced from. Fixing this (see https://github.com/oracle/graal/pull/2880/commits/bb94dcb04e5ea02d29ee92a7efff6b35466d9786) brings the binary size down to 18MBs.

  | master | this branch
-- | -- | --
-H:+OmitInlinedMethodDebugLineInfo | ~82s / 14MB | ~82s / 14MB
-H:-OmitInlinedMethodDebugLineInfo | ~90s / 17MB | ~130s / 18MB

I am keeping this PR in draft mode as there are more optimizations from @adinn that I haven't merged in yet.